requirements.txt
Django
djangorestframework
django-rest-auth